+++ components/AuthGuard.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { usePathname, useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 import { AppSidebar } from "@/components/sidebar";
 import { ProductSidebar } from "@/components/product-sidebar";
 import { ServiceSidebar } from "@/components/service-sidebar";
